Rename package from vortran to vortran-lbl for PyPI release
- Rename package: vortran -> vortran_lbl (directory and imports)
- Add LBL BSD 3-Clause license
- Switch build system from setuptools to uv
- Update all imports in tests and examples
- Update README examples to use vortran_lbl
- Update CI workflow to test vortran_lbl
