+++ .github/workflows/test.yml
@@ -31,4 +31,4 @@ jobs:
 
     - name: Run tests
       run: |
-        pytest --cov=src/vortran --cov-report=term-missing
+        pytest --cov=src/vortran_lbl --cov-report=term-missing
