+++ src/vortran_lbl/laser.py
